[level 1] Title: 정수 내림차순으로 배치하기, Time: 0.05 ms, Memory: 80.4 MB -BaekjoonHub

+++ 프로그래머스/lv1/12933. 정수 내림차순으로 배치하기/정수 내림차순으로 배치하기.java
@@ -15,7 +15,7 @@ class Solution {
         }
         
         String nStr = new String(nChar);
-        long answer = Long.parseLong(nStr);
+        long answer = Integer.parseInt(nStr);
         return answer;
     }
 }
